Refactor budget management: update Budget and BudgetStatus models to use Int for amounts, enhance BudgetCard with event handling for budget updates, and implement SettingViewModel for state management.

+++ app/src/main/java/com/example/subsense/core/data/model/BudgetStatus.kt
@@ -2,8 +2,8 @@ package com.example.subsense.core.data.model
 
 data class BudgetStatus(
     val category: ExpenseCategory,
-    val spent: Long,
-    val limit: Long,
+    val spent: Int,
+    val limit: Int,
     val percentage: Float,
     val isOverBudget: Boolean
 )
